requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
nltk
wordcloud
streamlit
newspaper3k
shap
joblib
